refactor(session-replay-browser): move beacon to beacon-transport at the top level
